+++ src/tests/test_openai_record.py
@@ -4,21 +4,22 @@ Test the OpenAI generator endpoints
 
 import shutil
 import tempfile
-from aoai_simulated_api.models import (
-    Config,
-    LatencyConfig,
-    ChatCompletionLatency,
-    CompletionLatency,
-    EmbeddingLatency,
-)
 
-from openai import AzureOpenAI, InternalServerError
+from openai import AzureOpenAI, InternalServerError, RateLimitError
 import pytest
 from pytest_httpserver import HTTPServer
 
 from .test_uvicorn_server import UvicornTestServer
 
 from aoai_simulated_api.record_replay.handler import get_default_forwarders
+from aoai_simulated_api.models import (
+    Config,
+    LatencyConfig,
+    ChatCompletionLatency,
+    CompletionLatency,
+    EmbeddingLatency,
+    OpenAIDeployment,
+)
 
 
 class TempDirectory:
@@ -90,6 +91,9 @@ def _get_replay_config(recording_path: str) -> Config:
             LATENCY_OPENAI_EMBEDDINGS_STD_DEV=0.1,
         ),
     )
+    config.openai_deployments = {
+        "low_limit": OpenAIDeployment(name="low_limit", model="gpt-3.5-turbo", tokens_per_minute=64 * 6)
+    }
     return config
 
 
@@ -109,7 +113,6 @@ async def test_openai_record_replay_completion(httpserver: HTTPServer):
     ).respond_with_data(
         '{"id":"cmpl-95FbXadIqJEMZZ1Rl0chTcKRxk2ez","object":"text_completion","created":1711038651,"model":"gpt-35-turbo","prompt_filter_results":[{"prompt_index":0,"content_filter_results":{"hate":{"filtered":false,"severity":"safe"},"self_harm":{"filtered":false,"severity":"safe"},"sexual":{"filtered":false,"severity":"safe"},"violence":{"filtered":false,"severity":"safe"}}}],"choices":[{"text":"This is a test","index":0,"finish_reason":"length","logprobs":null,"content_filter_results":{"hate":{"filtered":false,"severity":"safe"},"self_harm":{"filtered":false,"severity":"safe"},"sexual":{"filtered":false,"severity":"safe"},"violence":{"filtered":false,"severity":"safe"}}}],"usage":{"prompt_tokens":7,"completion_tokens":50,"total_tokens":57}}\n'
     )
-    httpserver.expect_request("/hello").respond_with_data("Hello, world!")
 
     with TempDirectory() as temp_dir:
         # set up simulated API in record mode
@@ -160,3 +163,67 @@ async def test_openai_record_replay_completion(httpserver: HTTPServer):
                 assert False, "Expected request to fail for non-recorded prompt"
             except InternalServerError as e:
                 assert e.status_code == 500
+
+
+@pytest.mark.asyncio
+async def test_openai_record_replay_completion_limit_reached(httpserver: HTTPServer):
+    """
+    Ensure we can call the completions endpoint multiple times using recorded responses to trigger rate-limiting
+    """
+
+    # Set up pytest-httpserver to provide an endpoint for the simulator
+    # to call in record mode
+    httpserver.expect_request(
+        uri="/openai/deployments/low_limit/completions",
+        query_string="api-version=2023-12-01-preview",
+        method="POST",
+    ).respond_with_data(
+        '{"id":"cmpl-95FbXadIqJEMZZ1Rl0chTcKRxk2ez","object":"text_completion","created":1711038651,"model":"gpt-35-turbo","prompt_filter_results":[{"prompt_index":0,"content_filter_results":{"hate":{"filtered":false,"severity":"safe"},"self_harm":{"filtered":false,"severity":"safe"},"sexual":{"filtered":false,"severity":"safe"},"violence":{"filtered":false,"severity":"safe"}}}],"choices":[{"text":"This is a test","index":0,"finish_reason":"length","logprobs":null,"content_filter_results":{"hate":{"filtered":false,"severity":"safe"},"self_harm":{"filtered":false,"severity":"safe"},"sexual":{"filtered":false,"severity":"safe"},"violence":{"filtered":false,"severity":"safe"}}}],"usage":{"prompt_tokens":7,"completion_tokens":50,"total_tokens":57}}\n'
+    )
+
+    with TempDirectory() as temp_dir:
+        # set up simulated API in record mode
+        config = _get_record_config(httpserver, temp_dir.path)
+        server = UvicornTestServer(config)
+        with server.run_in_thread():
+            aoai_client = AzureOpenAI(
+                api_key=API_KEY,
+                api_version="2023-12-01-preview",
+                azure_endpoint="http://localhost:8001",
+                max_retries=0,
+            )
+            prompt = "This is a test prompt"
+            response = aoai_client.completions.create(model="low_limit", prompt=prompt, max_tokens=50)
+
+            assert len(response.choices) == 1
+            assert response.choices[0].text == "This is a test"
+
+        # Undo httpserver config to ensure there isn't an endpoint to forward to
+        # when testing in replay mode
+        httpserver.clear_all_handlers()
+
+        # set up simulated API in replay mode (using the recording from above)
+        config = _get_replay_config(temp_dir.path)
+        server = UvicornTestServer(config)
+        with server.run_in_thread():
+            aoai_client = AzureOpenAI(
+                api_key=API_KEY,
+                api_version="2023-12-01-preview",
+                azure_endpoint="http://localhost:8001",
+                max_retries=0,
+            )
+            prompt = "This is a test prompt"
+            response = aoai_client.completions.create(model="low_limit", prompt=prompt, max_tokens=50)
+
+            assert len(response.choices) == 1
+            assert response.choices[0].text == "This is a test"
+
+            try:
+                aoai_client.completions.create(model="low_limit", prompt=prompt, max_tokens=50)
+                assert False, "Expect to be rate-limited"
+            except RateLimitError as e:
+                assert e.status_code == 429
+                assert (
+                    e.message
+                    == "Error code: 429 - {'error': {'code': '429', 'message': 'Requests to the OpenAI API Simulator have exceeded call rate limit. Please retry after 10 seconds.'}}"
+                )